pyproject.toml
[tool.poetry]
name = "anxiousbot"
version = "0.0.1"
description = "trading without patience"
authors = ["Filipe C Menezes <fcmenezes87@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.1"
ccxt = "^4.3.54"
python-json-logger = "^2.0.7"
python-telegram-bot = "^21.3"
pymemcache = "^4.0.0"
uvloop = "^0.19.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.4.2"
tabulate = "^0.9.0"
boto3 = "^1.34.137"
botocore = "^1.34.137"
selenium = "^4.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
format = "anxiousbot.scripts.format:_main"
debug = "anxiousbot.scripts.debug:_main"
main = "anxiousbot.main:_main"
marketcap = "anxiousbot.scripts.marketcap:_main"
config = "anxiousbot.scripts.config:_main"
test_bot = "anxiousbot.scripts.test_bot:_main"
